Fixed subspace quality policies
Added a maximum quality threshold at around 1e-4, so that the dynamic one wouldn't
cause the partial solver to give up finding a better subspace too early.
Also, if no state is found in the energy window, just give up and go to the next site.
This happens early on at the edges, where the energy spectrum might have large
jumps across the energy window of interest. To avoid "skating" in energy,
simply go to the next site, where the local hamiltonian might offer an state within
the energy window.


Former-commit-id: b435d15af3881778744c23c49cd98d98710f1f53