Fixed projection bug and cleaned subspace matrix build
At the center position I set the mps matrix  A = A * C = AC, by convention, where C is the center bond matrix (i.e., I absorb the schmidt values into A).
During projection I use a the "bare" A-matrix (without C) to tie off the edges. The bug was that I was using AC.  This appeared when I changed gamma-lambda -> A-B, but with solved I it looks like the change is completed.

Also, the subspace matrix code needed some freshening up


Former-commit-id: c5a94390e5519cfff7652851c284400ebcf67c06