Add support for shifting MPO energy before running the LBFGS in DIRECT mode
When we use "reduced-energy mpo's", the current energy per site is subtracted from all the MPO's at the beginning of every iteration.
The subtracted number is called "energy_reduced".
Directly after the energy subtraction, the target energy is exactly zero, and Var H = <(H-Er)²>
However, after some steps in the same iteration, the optimizer may have found a state with slightly different energy, so the target energy is actually 0 + dE.
We can obtain the shift amount dE = <H-Er>, which we call "eigval", i.e. the eigenvalue of the operator <H-Er>.
Then, technically Var H = <(H-E+|dE|)²>, and if unaccounted for, we may get Var H < 0, which is a real pain since we are optimizing the logarithm of Var H.
Here we use functor->set_shift in order to account for the shifted energy and reach better precision.